moved upload stuff to it's own package
